requirements.txt
Flask
Pillow
keyboard
psutil
pyautogui
pyperclip
requests